requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# PDF & OCR
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.1.0
pypdf2==3.0.1
pdfplumber==0.10.0
# OpenCV 4.10+ provides prebuilt wheels for Python 3.12; keep in sync with numpy
opencv-python-headless==4.10.0.84
# NumPy 1.26.4 is the last 1.x that supports Python 3.12 and works well with many wheels
numpy==1.26.4

# Utilities
python-magic==0.4.27
aiofiles==23.2.1